Korjattu ongelma tyhjien taulukoiden ja ylimääräisten käännösten kanssa
